requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
clarabel==0.9.0
click==8.1.8
contourpy==1.1.1
cvxpy==1.5.2
cycler==0.12.1
datasets==3.1.0
dill==0.3.8
ecos==2.0.14
exceptiongroup==1.3.1
fastapi==0.122.0
filelock==3.16.1
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.2.3
gurobipy==11.0.3
hf-xet==1.1.3
huggingface-hub==0.33.0
idna==3.10
importlib_resources==6.4.5
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
language_tool_python==2.8.2
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.1
nltk==3.9.1
numpy==1.24.4
osqp==1.0.4
packaging==25.0
pandas==2.0.3
pillow==10.4.0
pluggy==1.5.0
propcache==0.2.0
pyarrow==17.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.1.4
pytest==8.3.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scipy==1.10.1
scs==3.2.7.post2
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
sympy==1.13.3
tokenizers==0.20.3
tomli==2.3.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
